test(e2e): Move paginate test to its own file
